frontend Settings: Organize SettingsCluster and SettingsClusters
Moved from one big file into separate files named after their
components to follow convention.

Mainly renamed from ClusterSettings because that aliased an
existing type, and the new name matches the URL.
