+++ frontend/src/components/App/Settings/index.tsx
@@ -1,5 +1,7 @@
 import Settings from './Settings';
 import SettingsButton from './SettingsButton';
+import SettingsCluster from './SettingsCluster';
+import SettingsClusters from './SettingsClusters';
 export * from './Settings';
 export default Settings;
-export { SettingsButton };
+export { SettingsButton, SettingsClusters, SettingsCluster };
